Code Test Results
Attached is my project. I'm new to GitHub so had some trouble uploading the result, until I realized I had to switch to my fork and send it there. In my solution, I've used a couple of tool classes the I tend to reuse in projects for validation and code-reuse. Hope that's not cheating. Ran out of time before finishing some of the features and CSS.